requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-SocketIO==5.3.6
Werkzeug==3.0.1
gunicorn==21.2.0
Flask-WTF==1.2.1
Flask-Limiter==3.5.1
Flask-Session>=0.5.0
eventlet>=0.33.0
Pillow
cryptography>=42.0.0
psycopg2-binary>=2.9.9

# Optional: PostgreSQL support
# Uncomment the line below if you plan to use PostgreSQL instead of SQLite
# psycopg2-binary==2.9.9